Added support for multiple tokens in the server
Today the authentication script installed in the ejabberd server just validates user against one token. The authentication script should be able to validate a user against multiple tokens. To do that, the script was modified to fetch all tokens from the database and check each token received to try validate the user.

refs #1248